github_chat/requirements.txt
streamlit
langchain
faiss-cpu
langchain-community
beautifulsoup4
gitpython
python-dotenv 